Port `get_zulip_users` usage to `ZulipClient`
